systems/hello_world/requirements.txt
boto3==1.34.88
git+https://github.com/resim-ai/resim-python-client@e23220f9e51f6255929d2463d86772de8ccc5f79
polling2==0.5.0
requests==2.31.0
resim-open-core
